- XOVER's no long require 2 calls to dbzfetch if you're running with extendeddbz turned on.
- Cleaned up a few more isxxx() calls and a few printf("%ld", int)'s

- Reworked HISgetent so that it makes a little bit more sense.  It looks
like the "flag" parameter was being used for about 3 different things and it
was _incredibly_ confusing.  It appears to work with storage api, but I'm
not convinced it works for traditional spool.  As a side note, the XPATH
command appears to be completely broken and probably should be removed.

+++ nnrpd/post.c
@@ -168,7 +168,7 @@ TrimSpaces(p)
 
     for (start = p; ISWHITE(*start); start++)
 	continue;
-    for (p = start + strlen(start); p > start && CTYPE(isspace, p[-1]); )
+    for (p = start + strlen(start); p > start && CTYPE(isspace, (int)p[-1]); )
 	*--p = '\0';
     return start;
 }
@@ -209,7 +209,7 @@ StripOffHeaders(article)
     for (p = article; ; ) {
 
 	/* See if it's a known header. */
-	c = CTYPE(islower, *p) ? toupper(*p) : *p;
+	c = CTYPE(islower, (int)*p) ? toupper(*p) : *p;
 	for (hp = Table; hp < ENDOF(Table); hp++)
 	    if (c == hp->Name[0]
 	     && p[hp->Size] == ':'
